[NO ISSUE] Solved general issues and missing assets (#131)
* Adjusted Idle animation and added shotgun sound

* added missing sound effects for minotaur, rat and wolf

* corrected idle animation when inmune

* corrected Shotgun rate of fire

* corrected shotgun shells spawn position

* splitted animations for machinegun and shotgun

* added shotgun animations

* Added score for every enemy

+++ ZombustersWindows/GameObjects/Avatar.cs
@@ -173,7 +173,7 @@ namespace ZombustersWindows
         }
 
 
-        public bool VerifyFire(float currentSec, int rateoffire)
+        public bool VerifyFire(float currentSec, float rateoffire)
         {
             if (VerifyFire(currentSec, LastShot, rateoffire))
             {
@@ -183,7 +183,7 @@ namespace ZombustersWindows
             return false;
         }
 
-        private static bool VerifyFire(float currentSec, float lastShot, int RateOfFire)
+        private static bool VerifyFire(float currentSec, float lastShot, float RateOfFire)
         {
             if ((currentSec - lastShot) > (1f / RateOfFire))
                 return true;
